fix(graphs): keep wire exports and parity tests in their owners
