+++ src/jacobian/math/graphs/enumeration/__init__.py
@@ -3,13 +3,11 @@
 from jacobian.math.graphs.enumeration._models import (
     ColorPairCost,
     ConnectedColoredGraphFamily,
-    ConnectedColoredGraphsRequest,
 )
 from jacobian.math.graphs.enumeration.operations import connected_colored_graphs
 
 __all__ = [
     "ColorPairCost",
     "ConnectedColoredGraphFamily",
-    "ConnectedColoredGraphsRequest",
     "connected_colored_graphs",
 ]
